+++ Runtime/ARSubsystems/EnvironmentProbeSubsystem/XREnvironmentProbe.cs
@@ -30,6 +30,37 @@ namespace UnityEngine.XR.ARSubsystems
         /// <param name="descriptor">The <see cref="XRTextureDescriptor"/> associated with the probe's cubemap.</param>
         /// <param name="trackingState">The <see cref="TrackingState"/> of the probe.</param>
         /// <param name="nativePtr">A native pointer associated with the probe's cubemap.</param>
+        /// <param name="parentId">The <see cref="TrackableId"/> of the parent of this tracked object.</param>
+        public XREnvironmentProbe(
+            TrackableId trackableId,
+            Vector3 scale,
+            Pose pose,
+            Vector3 size,
+            XRTextureDescriptor descriptor,
+            TrackingState trackingState,
+            IntPtr nativePtr,
+            TrackableId parentId)
+        {
+            m_TrackableId = trackableId;
+            m_Scale = scale;
+            m_Pose = pose;
+            m_Size = size;
+            m_TextureDescriptor = descriptor;
+            m_TrackingState = trackingState;
+            m_NativePtr = nativePtr;
+            m_ParentId = parentId;
+        }
+
+        /// <summary>
+        /// Constructs an <see cref="XREnvironmentProbe"/>.
+        /// </summary>
+        /// <param name="trackableId">The <see cref="TrackableId"/> associated with this tracked probe.</param>
+        /// <param name="scale">The scale of the associated probe.</param>
+        /// <param name="pose">The <c>Pose</c> associated with the probe.</param>
+        /// <param name="size">The size (rendering bounds) of the associated probe.</param>
+        /// <param name="descriptor">The <see cref="XRTextureDescriptor"/> associated with the probe's cubemap.</param>
+        /// <param name="trackingState">The <see cref="TrackingState"/> of the probe.</param>
+        /// <param name="nativePtr">A native pointer associated with the probe's cubemap.</param>
         public XREnvironmentProbe(
             TrackableId trackableId,
             Vector3 scale,
@@ -46,6 +77,7 @@ namespace UnityEngine.XR.ARSubsystems
             m_TextureDescriptor = descriptor;
             m_TrackingState = trackingState;
             m_NativePtr = nativePtr;
+            m_ParentId = defaultValue.parentId;
         }
 
         /// <summary>
@@ -145,6 +177,16 @@ namespace UnityEngine.XR.ARSubsystems
             private set => m_NativePtr = value;
         }
         IntPtr m_NativePtr;
+
+        /// <summary>
+        /// The <see cref="TrackableId"/> of the parent of this tracked object.
+        /// </summary>
+        public TrackableId parentId
+        {
+            get => m_ParentId;
+            private set => m_ParentId = value;
+        }
+        TrackableId m_ParentId;
 
         /// <summary>
         /// Tests for equality.